refactor: 刪除 VideoUpload 和 UploadArea 元件以簡化結構，並更新引用路徑以改善檔案組織。
