+++ src/components/pages/VideoUpload/UploadArea.tsx
