+++ src/pages/upload-video/index.tsx
@@ -1,4 +1,4 @@
-import VideoUpload from '@/components/VideoUpload';
+import VideoUpload from '@/components/pages/VideoUpload';
 import { useAuth } from '@/hooks/useAuth';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
